requirements.txt
pytest>=7.4,<9.0
requests>=2.31,<3.0
allure-pytest>=2.13,<3.0
jsonschema>=4.21,<5.0
